move mod to c file
